Rebuild qt5-base: add support for mysql/mariadb driver
Summary:
The package akonadi needs mariadb/mysql for sane defaults (sqlite
use, although in theory possible, is not recommended by the devs).
Akonadi needs the driver to be enabled in qt, in order to use mariadb as backend.

Test Plan:
- I builded and installed the package
- After installation I ran akonadictl start. Previously it wouldn't start,
missing the driver for mysql. Now it starts and you can check that keeps running
(akonadictl status).

Reviewers: #triage_team, sunnyflunk

Reviewed By: #triage_team, sunnyflunk

Subscribers: ikey, JoshStrobl, sunnyflunk

Differential Revision: https://dev.solus-project.com/D205